Restructure on_drag, add missing reset to drag_over_id
check_hover is now done first, so everything now resides in one big "if drag_entity". Also, movement code was moved to the top and thus unified for being_created.

In the local function drag_over_id the hover_entity.temp_value was not reset when the mouse left the right area of an ID/Entity, that is fixed now.